[Enhancement] Make tablet checker blockingly add tablet to scheduler(#27648) (#27798)
Fixes SR-19074

1. Make `TabletChecker` blockingly add tablet to `TabletScheduler`, this
is to avoid meaningless loop and also avoid the situation that tablets
which have failed reparing for many times continuously added to pending
queue, and tablet which could be repaired successfully cannot get the
    chance to be added to scheduling queue.
2. Adjust the default value of `tablet_sched_max_balancing_tablets` and
`tablet_sched_max_scheduling_tablets` configurations, because we always
need to tell the user to increase those values.

This is a backport from pr #27648.

Signed-off-by: Dejun Xia <xiadejun@starrocks.com>