[Feature] add HTTP API /api/v2/warehouses (#27567)
## API
### Request
- URI: `API /api/v2/warehouses`
- Method: `GET`

### Successful Response
- Code: 200
- Body:
```json
{
    "status": "OK",
    "msg": "Success",
    "result": {
        "warehouses": [
          {
            "warehouse": "default",
            "num_unfinished_query_jobs": 1,
            "num_unfinished_load_jobs": 2,
            "num_unfinished_backup_jobs": 3,
            "num_unfinished_restore_jobs": 4,
            "last_finished_job_timestamp_ms": 1689652900112
          },
          {
            "warehouse": "test_warehouse",
            "num_unfinished_query_jobs": 1,
            "num_unfinished_load_jobs": 2,
            "num_unfinished_backup_jobs": 3,
            "num_unfinished_restore_jobs": 4,
            "last_finished_job_timestamp_ms": 1689652900112
          },
        ]
    }
}
```

### Error Response
- Code: not 200/201/202/203
- Response:
```json
{
    "status": "Failed",
    "msg": "<error-msg>"
}
```

## Implementation
Note that only the jobs which is submitted by users themself are
recorded in the above metrics.

### `WarehouseMetricMgr`
The number of unfinished query, backup, and restore jobs and last
finished timestamp are recorded by the metric counter. When the job is
starting, the counter++; when the job is finished, the counter -- and
record the finished timestamp.

### `WarehouseLoadInfoBuilder`
The number of unfinished load jobs computes according to `idToJobs` in
`LoadMgr`, `RoutineLoadMgr`, and `StreamLoadMgr`.

### All to One
The HTTP handler `WarehouseAction` fetches the information from all the
other FEs via the RPC `getWarehouses`, and will ignore all the failed
RPC response.

### Limitation
- The backup, restore, spark load, routine load, stream load job don't
support switch the current warehouse, which means their warehouse is
always `default_warehouse`.
- All the query and load jobs using `_statistic_`, which is the internal
statistic table, are considered as not user-queries.

---------

Signed-off-by: ZiheLiu <ziheliu1024@gmail.com>
(cherry picked from commit 1f5c50a57ed3f33a797cd4dd3b242d3af63fd241)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/http/rest/RestBaseAction.java
#	fe/fe-core/src/main/java/com/starrocks/qe/StmtExecutor.java