[Feature] add HTTP API /api/v2/warehouses (#27567)
- URI: `API /api/v2/warehouses`
- Method: `GET`

- Code: 200
- Body:
```json
{
    "status": "OK",
    "msg": "Success",
    "result": {
        "warehouses": [
          {
            "warehouse": "default",
            "num_unfinished_query_jobs": 1,
            "num_unfinished_load_jobs": 2,
            "num_unfinished_backup_jobs": 3,
            "num_unfinished_restore_jobs": 4,
            "last_finished_job_timestamp_ms": 1689652900112
          },
          {
            "warehouse": "test_warehouse",
            "num_unfinished_query_jobs": 1,
            "num_unfinished_load_jobs": 2,
            "num_unfinished_backup_jobs": 3,
            "num_unfinished_restore_jobs": 4,
            "last_finished_job_timestamp_ms": 1689652900112
          },
        ]
    }
}
```

- Code: not 200/201/202/203
- Response:
```json
{
    "status": "Failed",
    "msg": "<error-msg>"
}
```

Note that only the jobs which is submitted by users themself are
recorded in the above metrics.

The number of unfinished query, backup, and restore jobs and last
finished timestamp are recorded by the metric counter. When the job is
starting, the counter++; when the job is finished, the counter -- and
record the finished timestamp.

The number of unfinished load jobs computes according to `idToJobs` in
`LoadMgr`, `RoutineLoadMgr`, and `StreamLoadMgr`.

The HTTP handler `WarehouseAction` fetches the information from all the
other FEs via the RPC `getWarehouses`, and will ignore all the failed
RPC response.

- The backup, restore, spark load, routine load, stream load job don't
support switch the current warehouse, which means their warehouse is
always `default_warehouse`.
- All the query and load jobs using `_statistic_`, which is the internal
statistic table, are considered as not user-queries.

---------

Signed-off-by: ZiheLiu <ziheliu1024@gmail.com>