[BugFix] Choose the valid timestamp bits to compose the data cache key.
Before, we only use the hign 32 bits of timestamp, which only contains 9
valid bits. And many valid bits are ignored. So it can't reflect the file
modification in a timely manner.

In this PR, we only ignore the lowest 9 bits, and store the other 32 bits
to represent the second timestamp.

Signed-off-by: Gavin <yangguansuo@starrocks.com>