[BugFix] Disable local passthrough for global merger (#28155)
There is a clear conflict between local paththrough optimization and
global merger which requires the correct ordering of the data
transmition.

On the one hand, `SinkBuffer` and `PipelineSenderQueue` together
maintain a sliding window for the ordering of the data packets, which
requires the first sent packet must be processed first.

On the other hand, the local passthrough may cause a silly syndrome
where the first sent packet may potentially be processed later. which is
clearly defined in the pr #20550

And to a conclusion, we must disable the local passthrough optimization
when it comes to global merger.

## How to reproduce this problem

```sql
SELECT SUM(wv) FROM (
    SELECT max(fn_int) OVER (ORDER BY fn_int ROWS BETWEEN 2 PRECEDING AND 2 PRECEDING) AS wv FROM t_large
) tmp
```

the table `t_large` has 100M rows. and this sql may be blocked forever
at a relatively low ratio.

## What type of PR is this:
- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5
  - [ ] 2.4

Signed-off-by: liuyehcf <1559500551@qq.com>
(cherry picked from commit 99df623f972be0bf056cdfcf2c4f1eceee222b1f)