[BugFix] Release lock early in the drop tablet (#28058)
Drop tablet and release the lock quickly. The drop table will capture
the tablets_shard lock.
But there are running apply tasks on the tablet. So the tablets_shard
lock will be stuck on the apply thread.
The following thread are also stuck when try to get tablets_shard lock.