[BugFix] Delete unexpected rows when auto increment column is key (#28384)
Problem:
If the pk table with auto increment key run a delete txn without specifying
auto increment column, the increment column col value for the delete rows
will not been inited. And it will cause the problem that the un-inited value
for the col may hits some rows in current table and delete the row unexpectly.

Solution:
Assign 0 for the deleted row in auto increment col

Signed-off-by: srlch <linzichao@starrocks.com>