[BugFix][CherryPick] fix outer join and anti join rewrite bug for 2.5 (#28028) (#28466)
Fixes #27163

The bug has the following causes:
1. The invalid use of equivalence in outer join and anti join, which
leads to the invalid use of output columns. The join on predicate A.a =
B.b for outer join and anti join can not be used to construct
equivalence class.
2. the compatibility check should be more strict for outer and anti
join, which should not consider the pushdown predicates derived from
equivalence class. eg: the left outer join should make sure that the
predicates of right child tree of query and mv should be equal. it means
that the compensation predicates should not come from right child of
left outer join.
3. when clone equivalence class, should use the same
Set<ColumnRefOperator> for the columns in the same ec.

fixed by
1. constructing eq without column equal predicates in outer and anti
join
2. add more strict compatibility check for outer and anti join,
excluding pushdown ec derived predicates
3. fix equivalence class clone bug.

Signed-off-by: ABingHuang <codekhuang@163.com>