[BugFix] Support set tablet state in script engine to fix tablet in NOT_READAY state in tablet meta caused by #21223 (#28736)
Fixes #28735

Because of #21223, lots of tablets may already in NOT_READY state and
cannot do compaction.
Although #21224 fixes this bug, early versions without the fix may
already persist the bad state, which need to be updated. This PR addes a
method to set tablet states in script engine, so providing a way for
developer/SA to use scripts to fix error tablets. Example:

```
admin execute on 10004 '
for (info in StorageEngine.get_tablet_infos(-1, -1)) {
    if (info.state == 0) {
        var t = StorageEngine.get_tablet(info.tablet_id)
        if (t != null) {
            t.set_tablet_state_as_int(0)
            t.save_meta()
            System.print("fix table %(info.table_id) tablet %(info.tablet_id)")
        }
    }
}
';

```

Signed-off-by: Binglin Chang <decstery@gmail.com>