[BugFix] Getting analyzing error when hit the expression rewrite for generated column (#29247)
Problem:
If the expression in query hits the rewriting to generated column, it will get analyzing error
if using a alias for the table. Because the default table name for slotRef in generated column
expression will be the original one if the column is add by schema change rather than create table stmt.

Solution:
keep the expression unanalyze when schema change.

Signed-off-by: srlch <linzichao@starrocks.com>