[BugFix] Restore primary key table cause meta inconsistent when restart BE(#30129) (#30135)
Problem:
This problem is introduced by commit cfba8973f4718c0a4f2d6846d2bdaa473ba0be49
In this pr, we create a new tablet meta with snapshot tablet_schema to avoid the
crash if the backup table has been done schema change.

But the problem is that, the _updates in new tablet meta is nullptr. It forget to
reset it to tablet->updates(). So when remove expired versions for primary key table,
tablet_meta.save_meta() will save the wrong meta data for this tablet. When we restart
be, it will cause the meta inconsistent problem.

Solution:
reset the tablet_schema directly for the tablet_meta

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit 1c0a51347192a97973f484ede94ef4f2ec9be431)