[BugFix] Fix some bug of light schema change (#30485)
This pr fixes some bugs:

This pr([Feature] Support light schema change of adding and dropping columns #26246) supports light schema change and we add tablet schema in rowset meta. But in the course of multiple conflict resolution, the associated logic is lost. In addition, this pr change the location num of tablet_schema in RowsetMetaPB for better compatibility.
When we create a column iterator in segment, if we provide tablet schema in read options, we should use the given tablet schema in create column iterator.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>