[BugFix] generated column can not resloved in View(#31514)
Problem:
In visitSelect, generated column refs will be set in analyzeState using the refs in TableRelation.
It is unnecessnary to set for ViewRelation because the rewrite and be done throught normal TableRelation
and if we set it for ViewRelation, plan will get resloved error cause by the different context
between normal table and view.

Solution:
skip to setGeneratedExprToColumnRef for the ViewRelation

Signed-off-by: srlch <linzichao@starrocks.com>