[BugFix] Fix segment size in metadata cache
When loading a segment, only the footer and column meta are loaded,
and indexes are loaded lazily when they are used.
So segment memory usage is inaccurate when the segment is loading
and inserted into the metadata cache.
Therefore, the segment memory usage needs to be updated when
the indexes are loading.

Signed-off-by: wyb <wybb86@gmail.com>