[Refactor] Replan the insert to get rid of the metadata inconsistent between plan() and beginTransaction() (#31351) (#31584)
The metadata may be changed between plan() and beginTransaction(). When to beginTransaction(), the plan is not ready and the tablet is dropped by balanced. So we need to get txn first, to make the dropping tablet procedure wait.

(cherry picked from commit 0b35a46a9cbc676a5154d854a806e9d98d78aed3)