[Enhancement] Optimze `cosine_similarity` by avx2 (#31906)
The dimensions of the vectors are generally very high when performing brute-force vector retrieval calculations. In our online scenarios, there are at least 800 dimensions, and the use of SIMD instructions has a clear optimization effect. The optimization method is similar to: https://github.com/nmslib/hnswlib/blob/master/hnswlib/space_ip.h

The results are consistent with the test case results in the previous unit test：