[Enhancement] autovacuum withou list meta directory
Executing autovacuum tasks no longer performs a list
of tablet metadata. list tablet metadata is a costly
operation, and when there is a large amount of tablet
metadata (more than a few million), it takes a very
long time for the results to be returned, affecting
the speed of the vacuum.

Note that the autovacuum task still lists txn logs, which
normally doesn't have too many files and shouldn't take
too long, so it hasn't been optimized away for now.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>