[Enhancement] Enforce const and immutability in cache management
Changed a variety of variables and function returns to const to
enforce immutability in the cache management classes. Adjusted
some function calls and related logic to match these changes.
This will prevent the accidental modification of these data,
improving overall stability. Changed the shared_ptr in the
tablet metadata to const to emphasize the immutability of the
metadata after initial creation.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>