[BugFix] Table disable light schema change throws no global dict in meta scan (#33182)
After we enable light schema change, the tablet schema in BE maybe not the latest. So we need to regenerate tablet schema according to FE and we will generate a column unique id for each column in FE. But if a table created in old version, the column unique id in FE is default value(-1) and BE will assign the unique id to each column. So when we do meta scan and we generate the tablet schema according to FE, the column unique id is -1. And when we create column iterator, we can not find the right column iterator according to the unique id and it will throw no global dict exception when we do meta scan.

The solution is when we do meta scan for the table disable light schema change, we still use tablet schema in BE.