[BugFix] Get the inconsistent backup snapshot cause by the concurrent between data loading and backup (#33426)
Problem:
If the data loading and backup executing concurrently, the backuped snapshot for table metadata will
get inconsistent partition verison info that version + 2 = nextversion, but the correct version info
should be version + 1 = nextversion. If we restore such snapshot, we will get the table and it is can
not be loading data any more for primary key table.

Solution:
check the version info when backuping the table meta info

Signed-off-by: srlch <linzichao@starrocks.com>