[BugFix][Cherry-pick][Branch-3.1] BE crash in ASAN mode when we do column mode partial update for primary key table which separate primary keys and sort keys (#31219) (#33483)
When we do a column partial update for the primary key table which separates primary keys and sort keys, we will create a partial tablet schema for the rowset writer. However, the sort key columns may not exist in the partial tablet schema and the partial tablet schema will keep a wrong sort key idxes and short key column num. So BE will crash in ASAN mode. However, the sort_key_idxes and short_key_column_num in partial tablet schema are not important because the update segment file does not depend on it and the update segment file will be rewritten to a col file after apply. So we should modify the sort key idxes and short key column num in column mode partial update.