[BugFix] Partial update may lost some column's data (#33846)
If we update the tablet schema during a partial update(add a column), the newly added column data may be lost in the full segment due to the error logic in the check conflict.

The following is the origin code:
```
    if (latest_applied_version == _partial_update_states[segment_id].read_version &&
        _partial_update_states[segment_id].schema_version >= tablet_schema->schema_version()) {
        return Status::OK();
    }

    if (_partial_update_states[segment_id].schema_version >= tablet_schema->schema_version()) {
        Status st = _rebuild_partial_update_states(tablet, rowset, rowset_id, segment_id, tablet_schema);
        return st;
    }

```
If we update the schema during the partial update and find the tablet schema version is newer, we should rebuild  `RowsetUpdateState` because we may lose some columns. But in the logic, as shown above, we don't rebuild it.

e.g
1. The table schema is k1,v1,v2 before partial update, schema version is 0. We will prepare the data of column v2 in order to append it to the full rowset.
2. Add a new column v3 before apply and the tablet schema is k1,v1,v2,v3 and the schema version is 1.
3. During application, we should rebuild `RowsetUpdateState` and prepare the data of column v3. However, according to the logic above, the data of column v3 is lost.

(cherry picked from commit 6a85539cb1a814fe366a07bed11791a3803a72b4)