[Enhancement] Improve aborting transactions in shared data mode (#34185)
This patch contains the following changes:
1. Enhance transaction abort handling by including the ability to cancel active tasks.
2. Reduce the number of abort transaction tasks submitted to the thread pool to avoid filling up the thread pool queue.

In the previous implementation, if a transaction is aborted by FE because of timeout, it will send abortTxn() request to BE, and BE will try to read the txn log and clear the files recorded in the txn log, but no files will be deleted because the ingestion task is still in progress at this time, and the txn log is not generated. When the ingestion task on the BE is finished, the loadTxnCommit() request is sent to the FE, at this time, the commit will fail because the transaction is already in the aborted state, however, the BE will not clean up the files that have been written after the commit fails, so garbage data files are left on the object store.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>
(cherry picked from commit d103dc9e06610ab34155605405e7a937d0e4794c)

# Conflicts:
#	be/src/runtime/load_channel_mgr.h
#	be/src/service/service_be/lake_service.cpp
#	be/src/storage/lake/async_delta_writer.cpp
#	be/test/service/lake_service_test.cpp