[Enhancement] Refactor the create table logic to reduce the I/O usage. Currently, the create table will be splited multiple create_tablet tasks. Every create_tablet task will need two I/O operations, and always hold the lock, per RPC per tablet. The procedure is as followes. ``` 1. hold the tablet_shard lock      (hold the lock always) 2. create the tablet dir           (First I/O) 3. create tablet meta              (Second I/O) 4. add the tablet to the tablet_map 5. report the status to FE ```
If the concurrency of create table is high, the number of create_tablet tasks
may be very high. It will be stuck on the disk I/O and the lock.
So I batch the I/O operation for table granularity. Every table will cost
constant four I/O operations. As well the I/O opeartions is moved out of the lock.
Finally, it only have one RPC to response the status.
The new operation is as follows.
```
1. generate tablet dirs
2. save_create_txn(TXN_PREPARED)  (First I/O)
3. mkdir tablet dirs
4. sync tablet dirs               (Second I/O)
5. create tablet meta             (Third I/O)
6. save_create_txn(TXN_COMMITTED) (Four I/O)
7. hold the tablet_shard lock     (hold lock only for a while)
8. add the tablet to the tablet_map
9. report the status to FE
```

To complete the optimization, FE/BE use transaction mechanism
to guarantee the exactly-once for create table.
The transaction is only one-phase, not two-phases. When the BE succeed
to create the table, the tablets will be added to the map. So we need
to remove the tablet when FE failed, partial BEs succeed.
```
1. FE send an abort request to BEs at best effort.
2. BE remove the timeout transactions through an daemon thread.
3. BE report the tablets list to FE, FE send an request to remove the extra tablets.
```

If the transaction failed or timeout, the FE side remove the transaction
and send abort request to BE. The BE side will remove the