[Enhancement][1/n] Get tablet schema from a specific version of tablet metadata (#33466)
The goal of this patch is to ensure the obtained tablet schema is consistent
with the data being queried. In fact, in the current implementation, the tablet
schema is consistent across all versions of the tablet metadata. However,
this invariant will be broken soon due to the upcoming fast schema change
implementation, where the tablet schema may differ across versions of the
tablet metadata. Therefore, this patch can be seen as part of the fast schema
change implementation.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>
(cherry picked from commit a8f422e70cb6da76026b3d07560383c88d5ea82c)