[BugFix] conflict between light schema change and migration (#34542)
Problem:
Suppose there are two path: old_path and new_path.
In v3.1, a tablet with id 100 migrate from old_path to new_path
The tablet meta in old_path will be set shutdown.

If the shutdown tablet meta does not been gced and upgrade in v3.2
It is possible that the tablet meta in old_path will be replaced
by the one in new_path which is running state.

This will cause problem because there are two tablet with the same
id 100 in two different path and both of them are running state.

Signed-off-by: srlch <linzichao@starrocks.com>