[Enhancement] Allow skipping the cleanup process for abort transactions
This PR mainly aims to reduce unnecessary HEAD OBJECT requests.

The abort transaction will attempt to read the txn log and delete the
files recorded in the txn log. Typically, the transactions that will
be aborted have no txn log, which means that many times the cleanup
process for abort transactions will not have any real effect. In the
current implementation, even reading a non-existent txn log will send
a HEAD OBJECT request to the object store, and when the abort
transaction involves a large number of tablets, a large number of
HEAD OBJECT requests will be sent, which may result in the qps limit
of the object store being exceeded, and also cause unnecessary billing.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>