[BugFix] Fix data race of Rowset::load and Rowset::close between compaction and adding generated column (#35362)
There is a data race of Rowset::load and Rowset::close. The caller of the rowset->load will return successfully,
but will get an ROWSET_UNLOADED rowset later. An example is a duplicate key table that performs compaction and
adds a generated column:

Suppose we have a rowset r1 whose state is ROWSET_LOADED at first
Thread 1 (compaction) Thread 2 (adding generated column)

Schema Change Thread: Acquire rowset r1 while holding the tablet lock,
then release the lock to start performing link schema change and calculate
the value of the generated column。
Compaction Thread: Begin to compaction with the rowset r1 as one of input rowset.
Ready to commit compaction and execute rowset->close() for r1 with the rowset lock,
and get the reader reference count is 0.
Schema Change Thread: Execute rowset->load() for r1, because r1->state == ROWSET_LOADED, return
successfully with acquiring rowset lock.
Compaction Thread: clear the segment metadata for r1 in memory and set the state as ROWSET_UNLOADED
for r1.
Schema Change Thread: crash because the segment metadata in memory has been clear at this time.

The reason why a race condition occurs here is that:
In rowset close, the rowset lock will be obtained first, and then the reader reference count
snapshot of the rowset reader will be obtained. If the value of the reference count snapshot is 0,
the metadata cached in the memory of the segment will be cleared.

Just before clearing the segment metadata cache, if the thread switches to schema change.
Schema change requires executing rowset->load, but since rowset statue is ROWSET_LOADED at this time,
it will return directly without lock. At this time, the thread switches back to commit compaction and
clear the segment meta in memory. This will cause the segment metadata to be empty when the subsequent
thread switches back to the schema change and a crash occurs.

The pull request starts to acquire the rowset lock in rowset->load even in the ROWSET_LOADED state

Signed-off-by: srlch <linzichao@starrocks.com>