[BugFix] MV creation maybe failed (#35492)
After support fast schema evolution, the tablet schema in BE maybe not the latest. So we will check the column unique id when we create a new tablet(alter/rollup) to avoid serious issue. However, if we drop a column first and add a new column with same column name secondly, if the tablet schema in BE is not the latest, the check will failed because these two columns with same column name are not the same column. So we should first check schema version, but when we create a rollup, the schema version is not set which cause check failed.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>