[BugFix] Fix BE graceful exit (#35361)
Why I'm doing:

The graceful exit of the BE might fail due to incorrect resource destruction order.

What I'm doing:

Fix resource release order to make graceful exit work.

Fixes #issue

In the destruction of storage engine instance in BE, we will first set the instance to nullptr and release the other resource. But during destruction of tablet manager, we maybe access the storage engine instance which cause nullptr exception.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
Signed-off-by: Astralidea <astralidea@163.com>