[BugFix] Avoid being unable to select when partition_duration is set incorrectly
Why I'm doing:

Before this submission, when partition_duration was set incorrectly, selection
could not be performed and "Unsupported unit: Seconds" was returned directly.

What I'm doing:

This submission modifies this behavior: when partition_duration is set incorrectly,
we convert the datacache.partition_duration type from date to datetime to compare
with the partition range and catching the other exceptions (if any) during time comparison.

Signed-off-by: stonechen111 <chenshi@starrocks.com>