[BugFix] BE in ASAN mode may crash when ingestion and drop tablet concurrently (#33432)
When we drop a primary key tablet, we will delete the tablet meta include primary index cache, delete vector and so on.  But when the  ingestion and drop tablet run concurrently, there are some issues:
1. When we delete primary index, we will check only one thread holds it, otherwise BE will crash in ASAN mode. But the ingestion thread also hold the primary index when it preload `RowsetUpdateState`, so BE maybe crash in ASAN mode when drop tablet.
2. Ingestion maybe get `no delete vector found` exception because  delete vector maybe deleted due to drop tablet.

BTW, this pr is just a temporary solution and the root cause is when we drop primary key table, we will clean up the tablet metadata include primary index, delete vector, etc. and these metadata might still be accessed.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>