[Refactor] Add detail failed info when create table timeout (#36378)
Why I'm doing:

If table creation fails due to a timeout, it could be because the timeout duration is too short or due to some BE failures. The current response doesn't include the information about the failed BE, which makes it difficult to diagnose the cause of the timeout.

What I'm doing:

Add failed BE info into response.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
Signed-off-by: 鼎昕 <zhangdingxin.zdx@alibaba-inc.com>