[Enhancement] scale compaction score to a reasonable and user friendly range (#36534)
Why I'm doing:

The compaction score for pk tablet is based on bytes unit which is different from other model(based on file number), which is hard for user to understand and compare.

What I'm doing:

Scale the score to a range that is somewhat similar to file number * 10 (multiply by 10 to prevent lost all fractional part)

After this PR, the compaction score for PK tablet showed to user/dev will become much smaller.

Fixes #36468

Signed-off-by: Binglin Chang <decstery@gmail.com>
Signed-off-by: 鼎昕 <zhangdingxin.zdx@alibaba-inc.com>