[BugFix] Fix DataStreamRecvr profile use-after-free
Suppose a query is a short-circuited query, and the query execution is ready to finish on Fragment A of some BE, when some rpc from another BE arrives.
If Fragment A destructs. At this point SenderQueue access to the profile object will result in use-after-free. our DataStreamRecvr needs to satisfy closure.

Signed-off-by: stdpain <drfeng08@gmail.com>