[BugFix] create bitmap index failed if table already has another bitmap (#37116)
Why I'm doing:
If creating a Bitmap index for a table that already has a Bitmap index, it will fail.
This problem is introduced by pr(#36765) which introduced the meta data for GIN(inverted index).

This pr add a index id for every index. Bitmap index always has id = -1 and GIN has id id >= 0
In the process of CREATE INDEX, it will check the new added index' id is larger than the existed index'id
or not. This is make sense for GIN but not for BITMAP(always -1).
If we add two bitmap indexes consecutively, it will trigger a failure cause by (-1 > -1) is FALSE.

What I'm doing:
Check the id for GIN only

Signed-off-by: srlch <linzichao@starrocks.com>