[BugFix] Fix pindex read/write confilict when enable pindex rebuild in compaction (#37197)
Why I'm doing:

This bug is introduced by this pr(#36819), we will rebuild pindex if we pick all rowsets to do compaction. But in rebuild pindex, we try to reset pindex but doesn't hold the lock which may cause read/write conflict between compaction and data ingestion.

What I'm doing:

Hold the primary index lock when we reset primary index.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>