[Enhancement] Support to configure the number of http worker threads on FE (#37530)
Why I'm doing:
Http worker threads on FE are used to deal with http requests.  Some requests may do heavy operations which will block threads, such as transaction stream load commit, and other operations that need to get lock. As a result, http server can't deal with other requests.  Before optimizing these heavy operations, we first make the number of http worker threads to be configurable, and increase it as a candidate solution if the problem happens.

What I'm doing:
Add a configuration to adjust the number of http worker threads on FE

Signed-off-by: PengFei Li <lpengfei2016@gmail.com>