[BugFix] Should not add full row column in vertical compaction's rowset writer (#38385)
Why I'm doing:

There is a bug in segment writer's write path, when doing vertical compaction, it will mistakenly consider it a case to add a full row column in the input chunk, causing BE to crash.

What I'm doing:

Change the condition logic to rule out this case. Also change all "__row" usage to using a const var.

Fixes #38132

Signed-off-by: Binglin Chang <decstery@gmail.com>
Signed-off-by: 戴鑫铉 <xinxuan.dai@mihoyo.com>