[BugFix] Dictionary refresh maybe timeout if the dictionary object is huge (#39380)
Why I'm doing:
If we refresh a dictionary cache using a huge dictionary table, refresh timeout may
happen. The problem is that, we refresh 256MB chunk for a single chunk in a single
refresh task. Each refresh task is executed concurrently. Since this chunk is
relatively large, the respond time of a certain task's rpc request may be too long
and cause a timeout.

What I'm doing:
1. Make refresh cache rpc timeout configable.
2. Make chunk smaller when it used for refresh task. Helps reduce rpc response time.
3. Make the refresh task thread pool size configable. By controlling concurrency, speed
and task pressure can be balanced.
4. Unlimit the queue size of task thread pool. Because tasks are divided into smaller pieces,
the number of tasks may increase.

Signed-off-by: srlch <linzichao@starrocks.com>