[Enhancement] Make some operation type ignorable when replaying journal fails (#39091)
Why I'm doing:
Journal replay failure is the main reason for FE startup failure, but most Journal replay failures only affect the use of a small function. Therefore, it is necessary to sort out the severity of each replay failure and ignore the OperationType with a relatively small impact.

What I'm doing:
The replay failure of OperationType with @IgnorableOnReplayFailed annotation will be ignored when `Config. metadata_journal_ignore_replay_failure` is set to true.

Signed-off-by: gengjun-git <gengjun@starrocks.com>