[BugFix] Fix forward to self node bug (#39587)
Why I'm doing:
If an stmt that should be forwarded to the leader, such as "show frontends", is executed on the node that is being transferring to the leader, the stmt will be forwarded to the self node multiple times, and the result will be wrong.

The reason is that: the forward result is not processed in `ConnectProcessor.proxyExecute`.

What I'm doing:
Just return the forward result if stmt is also forwarded during execution in `ConnectProcessor.proxyExecute`.
Furthermore, in order to limit the forwarding times to a smaller value, there are 2 enhancements:
1. If the number of forwards is greater than 30, the forwarding chain is interrupted and an exception is reported.
2. If the current node is being transferred to the leader, the initForwardToLeaderState() method is blocked until the transfer is completed.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit d8bb8329299f0d96b18702dd006f3805498096fb)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/common/ErrorCode.java
#	fe/fe-core/src/main/java/com/starrocks/qe/ConnectContext.java
#	fe/fe-core/src/main/java/com/starrocks/qe/LeaderOpExecutor.java
#	fe/fe-core/src/main/java/com/starrocks/server/GlobalStateMgr.java
#	gensrc/thrift/FrontendService.thrift