[BugFix] Fixed the mem statistics problem of manual compaction (#41150)
Why I'm doing:
There may be deviation in memory statistics for each compaction task, so when finishing the task, the held MemTracker must be destructed so that errors will not accumulate. However, manul compaction for the primary key model and other model with enable_size_tiered_compaction_strategy=false, is actually using
global compaciton mem tracker, which is a global mem tracker and will not be destructed, which will lead to accumulation of deviation so the statistical compaction memory will continue to increase.

What I'm doing:
Create one child mem tracker for each compaction task
The code of manual compaction is added after fork, so remove the old invalid license.

Signed-off-by: trueeyu <lxhhust350@qq.com>
Signed-off-by: liqingzhong <liqingzhong@inspur.com>
Signed-off-by: qingzhongli <qingzhongli2018@gmail.com>