[BugFix] Replace the `rapidjson::StringRef` with `SetString` to avoid refrencing a temporary string pointer which cause `stack-use-after-return` memory issue in datacache http action. (#41663)
Why I'm doing:
We use rapidjson::StringRef in the handler of datacache action to construct the response content. However, it only reference an exist string pointer without copying its content. In our case, it reference a local string pointer, and may cause stack-use-after-return memory issue when the local string is freed.

What I'm doing:
Use the SetString to copy the string content and avoid this problem.

Signed-off-by: Gavin <yangguansuo@starrocks.com>