[BugFix] Add the missing closure ref count in multi_get (#42005)
Why I'm doing:

multi_get function is used to get the corresponding value
for the given key value in some primary key tablets. If the
tablet is not local, be will call a rpc to get the value
from the remote node.
The problem is closure->ref() is miss before call the rpc.
When the rpc has finished, --closure->_ref.
It will be deleted in this case, because the ref count = 0.
So we will get many unpredictable BE crash because use after
free problem.

What I'm doing:

Add a additional closure->ref() before call rpc.

Signed-off-by: srlch <linzichao@starrocks.com>