[BugFix] Fixe bug where tablet was not deleted from meta when removing disk path from be.conf (#41755)
## Why I'm doing:
A disk is removed from be's conf. The normal repair logic is as follows:
1. BE reports the full amount of tablets to FE. At this time, it will not bring along the tablets that have been removed from the disk.
2. FE compares the tablets reported by BE with the tablets recorded in its own metadata, and deletes replicas of tablets that do not exist in BE from the metadata.
3. Make up for the lost replicas based on the tablets on other BEs.

However, if there are too many tablets on a BE, the processing time of the second step is too long. As long as tablets are created or imported on the BE, which is controlled by reportVersion, the deletion will not be performed, resulting in incorrect metadata information being recorded on the FE side.

## What I'm doing:
Only check reportVersion when the disk is online, as there will be no tablet changes on an unavailable disk.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 805e22152ad7403ae0143700e173406887a10637)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/leader/ReportHandler.java