[BugFix] Reset replica minReadableVerion in restore
Why I'm doing:
Support we have a table with version 10 and minReadableVersion 5.
And we have a snapshot for this table with version 2. When we restore
this snapshot directly without dropping the original table, the replica's
minReadableVersion for it have not been reset. So we have the partition
with visibleVersion = 2 and minReadableVersion = 5 which is wrong.

What I'm doing:
reset the minReadableVersion in restore if necessary.

Signed-off-by: srlch <linzichao@starrocks.com>