[BugFix] Fix generated column become NULL after compaction in pk table
Why I'm doing:
Suppose we have a PK table t with c1, c2, c3 and execute SQL as following:
1. INSERT INTO t VALUES (1,1),(2,2),(3,3); rowset 1 with tablet_schema version 0
2. drop column c3;
3. INSERT INTO t VALUES (1,1),(2,2),(3,3); rowset 2 with tablet_schema version 1
4. add column c4 as c1 + c2;

After finishing adding c4, if compaction begin with rowset 1 and rowset 2, it will
use schema in rowset 2 to do compaction, because rowset 2 has greater schema version
than rowset 1. But the problem is, use schema(c1, c2) will miss the column c4, SegmentIterator
will not read the value of the generated column c4 and lead to be NULL after compaction.

This is a corner case about generated columns and will only affect generated columns and pk tables.

What I'm doing:
Reset the schema to the latest for all rowset in link_from

Signed-off-by: srlch <linzichao@starrocks.com>