[Enhancement] Change the parameter strict_mode to null_if_not_found of dict_mapping function(#40897)
The last optional parameter for dict_mapping is strict_mode(boolean), and its behavior
if following:
If strict_mode is True : return an exception if the key is not found in dictionary table.
If strict_mode is False(Default) : return null if the key is not found in dictionary table.

We change the strict_mode into null_if_not_found to make more clear for user to use this function,
and the behavior of null_if_not_found is following:
If null_if_not_found is True : return null if the key is not found in dictionary table.
If null_if_not_found is False(Default) : return an exception if the key is not found in dictionary table.

Signed-off-by: srlch <linzichao@starrocks.com>