[Feature](3/n) Support fast schema evolution in shared data mode
Before the fast schema evolution is implemented, the schema of a
MaterializedIndex will not change, and the materialized index ID
can be used as the unique identifier of its schema. In the
upcoming fast schema evolution, the schema of a materialized index
may change, and the materialized index id can no longer be used
as the unique identifier of the schema. This PR changes the use
of index id to schema id.

Signed-off-by: Alex Zhu <zhuming9011@gmail.com>