[BugFix] Fix disk metrics bug when partition name does not end with number. (#42723)
When the partition name does not end with a number, it is incorrect to trim the suffix number to get the device name. For example, the device name is `nvme0n1` and the partition name is `nvme0n1p1`, in this case the device name generated from the partition is `nvme0n1p` and using this name to find statistics from `/proc/diskstats` will fail.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 762ca68b6a23b32df312a353d5f30ea8952c600f)