[BugFix] Fix shutdown tablet meta leak cause migration fail.
Why I'm doing:
Suppose we have there different datadir called dir1 dir2 dir3.
One of the problematic execution timeline is following:
1. Create a tablet with id 123 on dir1, we call it tablet_123_dir1.

2. migrate it into dir2, tablet_123_dir1 will be added into shutdown_tablets
map but will not remove the meta data in rocksdb immediately. And we
re-create a tablet_123_dir2 on dir2, migration is done.

3. migrate it into dir3, tablet_123_dir2 will be try to added into shutdown_tablets.
But problem happen here: we actully can not add tablet_123_dir2 into shutdown_tablets
because we use std::map::emplace to add the tablet, but <123, tablet_123_dir1> has
already existed in this map, std::map::emplace will not rewrite <123, tablet_123_dir1>
into <123, tablet_123_dir2>. So we actully destory tablet_123_dir2 immediately without
remove its meta data. And the meta data for tablet_123_dir2 in rocksdb can not be removed
forever in this case.

4. migrate it back to dir2. Now we notice that there is tablet meta data in dir2 for tablet 123.
But we can not remove it this case, because tablet_123_dir2 is gone.

The core problem here is that shutdown_tablets is std::map and we use std::map::emplace to add
a new shutdown tablet. It may leak if the tablet id has been existed in this map.

What I'm doing:
Forcely remove the previous shutdown tablet with same tablet id if add a new one to it.

Signed-off-by: srlch <linzichao@starrocks.com>