[UT] Fix PseudoBackend report tablets bug (#43165)
Report_version must be set before setting the tablets to ensure that the tablet in FE will not be accidentally deleted in the following case:
t1: tablet1, tablet2, tablet3 are created on BE and current reportVersion is 3.
t2: tablet1, tablet2, tablet3 are set into request.tables.
t3: tablet4 is created and reportVersion changed to 4.
t4: 4 is set into request.report_version.
t5: The request is sent to the FE node and it is found that tablet4 is not in request.tablets
and request.report_version is equal to the latest report_version recorded by FE,
tablet4 metadata will be deleted from FE. Code: ReportHandler.deleteFromMeta

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 9a12ac91006cdd4ec181c5754e62bb171e7969c8)